fix: 지난 시간 예약 조회가 실패하는 버그 해결
