+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,7 +1,6 @@
 package roomescape.domain;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 
 public class Reservation {
 
@@ -12,7 +11,6 @@ public class Reservation {
     private final Theme theme;
 
     public Reservation(Long id, String name, LocalDate date, ReservationTime reservationTime, Theme theme) {
-        validateDateIsFuture(toLocalDateTime(date, reservationTime));
         this.id = id;
         this.name = name;
         this.date = date;
@@ -22,16 +20,6 @@ public class Reservation {
 
     public Reservation(String name, LocalDate date, ReservationTime reservationTime, Theme theme) {
         this(null, name, date, reservationTime, theme);
-    }
-
-    private void validateDateIsFuture(LocalDateTime localDateTime) {
-        if (localDateTime.isBefore(LocalDateTime.now())) {
-            throw new IllegalArgumentException("지나간 날짜와 시간에 대한 예약 생성은 불가능합니다.");
-        }
-    }
-
-    public LocalDateTime toLocalDateTime(LocalDate date, ReservationTime reservationTime) {
-        return LocalDateTime.of(date, reservationTime.getStartAt());
     }
 
     public boolean isSameReservation(Long id) {
