+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 class ReservationTest {
 
@@ -34,24 +34,5 @@ class ReservationTest {
                 new Theme("capyTheme", "capyDescription", "capyThumbnail"));
 
         assertThat(reservation1.isSameReservation(reservation2.getId())).isFalse();
-    }
-
-    @Test
-    @DisplayName("미래의 날짜와 시간에 대한 예약 생성이 가능하다.")
-    void checkReservationDateTimeIsFuture_Success() {
-        assertThatCode(() -> new Reservation(1L, "capy", LocalDate.now().plusDays(1L),
-                new ReservationTime(LocalTime.now()),
-                new Theme("capyTheme", "capyDescription", "capyThumbnail")))
-                .doesNotThrowAnyException();
-    }
-
-    @Test
-    @DisplayName("지나간 날짜와 시간에 대한 예약 생성시 예외가 발생한다.")
-    void checkReservationDateTimeIsFuture_Failure() {
-        assertThatThrownBy(() -> new Reservation(1L, "capy", LocalDate.now().minusDays(1L),
-                new ReservationTime(LocalTime.now()),
-                new Theme("capyTheme", "capyDescription", "capyThumbnail")))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("지나간 날짜와 시간에 대한 예약 생성은 불가능합니다.");
     }
 }
