+++ src/test/java/roomescape/service/reservation/ReservationCreateServiceTest.java
@@ -73,4 +73,19 @@ class ReservationCreateServiceTest {
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("해당 시간에 이미 예약된 테마입니다.");
     }
+
+    @Test
+    @DisplayName("지나간 날짜와 시간에 대한 예약 생성시 예외가 발생한다.")
+    void checkReservationDateTimeIsFuture_Failure() {
+        ReservationTime reservationTime = reservationTimeCreateService.createReservationTime(
+                new SaveReservationTimeRequest(LocalTime.now().plusHours(1L)));
+        Theme theme = themeCreateService.createTheme(
+                new SaveThemeRequest("capy", "caoyDescription", "caoyThumbnail"));
+        SaveReservationRequest request = new SaveReservationRequest(
+                "capy", LocalDate.now().minusDays(1L), reservationTime.getId(), theme.getId());
+
+        assertThatThrownBy(() -> reservationCreateService.createReservation(request))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessage("지나간 날짜와 시간에 대한 예약 생성은 불가능합니다.");
+    }
 }
